Finished 01B Wrangle
Tidied up comments, made sure everything runs. There is a possible issue with running the calculations in an Rmd file as the shp files are too large. Final function was struggling to write out every time, but it works fine on smaller shp files, so I assume this is a memory issue.